fix(ai): prevent migration of dynamic environment API key fallbacks into persistent storage
